requirements.txt
flask
gunicorn
python-dotenv
flask-wtf
pymongo
pymongo[srv]
flask-security
email_validator
passlib